updating task 0

+++ 0x14-bit_manipulation/0-binary_to_uint.c
@@ -9,7 +9,7 @@
  */
 unsigned int binary_to_uint(const char *b)
 {
-	int i, power = 1, number = 0;
+	unsigned long int i, power = 1, number = 0;
 
 	if (b == NULL)
 		return (0);
